Collision chokes on typed repository methods

+++ database/migrations/2018_08_08_100000_create_telescope_entries_table.php
@@ -8,17 +8,10 @@ return new class extends Migration {
     /**
      * Get the migration connection name.
      */
-    public function getConnection(): string
+    public function getConnection(): ?string
     {
-        return config()->string('telescope.storage.database.connection');
-    }
-
-    /**
-     * Determine if this migration should run.
-     */
-    public function shouldRun(): bool
-    {
-        return config()->boolean('telescope.enabled');
+        /** @phpstan-ignore-next-line return.type `nunomaduro/collision` and its test command fails with typed config repository methods */
+        return config('telescope.storage.database.connection');
     }
 
     /**
@@ -26,6 +19,10 @@ return new class extends Migration {
      */
     public function up(): void
     {
+        if (!config('telescope.enabled')) {
+            return;
+        }
+
         $schema = Schema::connection($this->getConnection());
 
         $schema->create('telescope_entries', function (Blueprint $table): void {
@@ -68,6 +65,10 @@ return new class extends Migration {
      */
     public function down(): void
     {
+        if (!config('telescope.enabled')) {
+            return;
+        }
+
         $schema = Schema::connection($this->getConnection());
 
         $schema->dropIfExists('telescope_entries_tags');
